edit user reasource

+++ app/Http/Controllers/Api/UserController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\Api\User\EditProfileRequest;
 use Illuminate\Support\Facades\Auth;
 use App\Traits\ApiTraits;
 use App\Traits\HelperTrait;
+use App\Http\Resources\Api\UserResource;
 
 class UserController extends Controller
 {
@@ -17,12 +18,15 @@ class UserController extends Controller
         $user = Auth::user();
         if($request->image){
             if($user->image){
-                unlink($user->image);
+                if (file_exists($user->image)){
+                    unlink($user->image);
+                }
             }
             $image = $this->uploadImages($request->image, "images/user/profile");
             $user->update(['image' => $image]);
         }
         $user->update($request->except('image'));
-        return $user;
+        return $this->responseJson(200, "Profile Updated Successfully", new UserResource($user));
+
     }
 }
